src/producer/requirements.txt
azure-eventhub
aiohttp
python-dotenv
